services/data-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
influxdb-client==1.40.0
paho-mqtt==1.6.1
redis==5.0.0
httpx==0.27.0
python-multipart==0.0.9
pydantic-settings==2.5.0
pytest==8.3.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
websockets==12.0
pandas==2.2.0
